streamlit/pyproject.toml
[project]
name = "text2sql-streamlit"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.coverage.run]
source = ["streamlit"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.streamlit/*",
    "app.py",  # Exclude UI layer
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
